Домашка по "Продолжение знакомства с SQL". Поправил замечания, пока кроме вызова листа с авами.

+++ src/main/java/ru/hogwarts/school/controller/StudentController.java
@@ -75,7 +75,7 @@ public class StudentController {
     }
 
     @GetMapping("/get-avg-age")
-    public Long getAvgOfStudentsAge() {
+    public Float getAvgOfStudentsAge() {
         return studentService.getAvgOfStudentsAge();
     }
